mqtt-device-ui/requirements.txt
Flask==2.2.5
kubernetes==26.1.0
paho-mqtt==1.6.1
